overlake: memory layout
U-Boot patches memory alyout in order to allow part of physical memory
to be used for preserved memory.

However, we also want to be able to use DTB from fitImage without relying
on U-Boot in order to be able to update DTB via kexec if needed.

Set the physical memory layout to be the same as what Overlake board has
minus the 2G of persistent memory.

Signed-off-by: Pavel Tatashin <pasha.tatashin@soleen.com>

TAG_MSFT: overlake