Merged PR 2232499: Bug 23780818: EV1/2 board resets when setting up encrypted partition
CRMU Page table memory is not reserved in Linux DTS. This was causing the Linux to overwrite the page tables and getting CRMU into bad state. This causes TMON to read a bad value and leading to Critical thermal shutdown and reboot.

Bug 23780818: EV1/2 board resets when setting up encrypted partition

Verification:
- I verified that "dd" stress does not reboot the system abnormally. I also booted overlake specialized image and let it stay there for few hours with dd running. Without the patch, Overlake specialized image used to reboot on that failing board.
- Shiping also regressed his scenarios with the patch
- Waiting for Simon to regress